agent-platform/backend/requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.5.2

# Database
sqlalchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.9

# LangChain & AI
langchain==0.3.7
langchain-core==0.3.15
langchain-openai==0.2.5
langchain-anthropic==0.2.4
langchain-chroma==0.1.4
chromadb==0.5.11

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12

# Utilities
python-dotenv==1.0.1
httpx==0.27.2
tenacity==9.0.0
aiofiles==24.1.0

# Development
pytest==8.3.3
pytest-asyncio==0.24.0
black==24.10.0
ruff==0.7.3
